Prefer jakarta apis over older javax apis. (#3069)
