+++ .github/workflows/maven.yml
@@ -38,6 +38,6 @@ jobs:
         restore-keys: |
           ${{ runner.os }}-maven-
     - name: Build with Maven
-      run: ./mvnw -V -B -ff -s .github/settings.xml install '-Dmaven.test.skip=true' '-Dmaven.javadoc.skip=true'
+      run: ./mvnw -Pjakarta-apis -V -B -ff -s .github/settings.xml install '-Dmaven.test.skip=true' '-Dmaven.javadoc.skip=true'
     - name: Run tests
-      run: ./mvnw -V -B -ff -s .github/settings.xml verify
+      run: ./mvnw -Pjakarta-apis -V -B -ff -s .github/settings.xml verify
