+++ appveyor.yml
@@ -17,8 +17,8 @@ cache:
 install:
 - cmd: 'java -version'
 build_script:
-- mvnw.cmd -V -B -ff install -Dmaven.test.skip=true -Dmaven.javadoc.skip=true -Dmaven.source.skip=true -Dmaven.deploy.skip=true
+- mvnw.cmd -Pjakarta-apis -V -B -ff install -Dmaven.test.skip=true -Dmaven.javadoc.skip=true -Dmaven.source.skip=true -Dmaven.deploy.skip=true
 test_script:
-- mvnw.cmd -V -B -ff verify -Dmaven.javadoc.skip=true -Dmaven.source.skip=true -Dmaven.deploy.skip=true
+- mvnw.cmd -Pjakarta-apis -V -B -ff verify -Dmaven.javadoc.skip=true -Dmaven.source.skip=true -Dmaven.deploy.skip=true
 matrix:
   fast_finish: true
