+++ dropwizard-jetty/pom.xml
@@ -18,22 +18,22 @@
             <artifactId>dropwizard-logging</artifactId>
         </dependency>
         <dependency>
-            <!-- this is required because jetty-server pom
-            put servlet-api into default scope when it should
-            always be provided scope so it doesn't get compiled.
-            without this there are security exceptions because
-            javax.servlet-api jar is not signed -->
-            <groupId>javax.servlet</groupId>
-            <artifactId>javax.servlet-api</artifactId>
-            <scope>provided</scope>
-        </dependency>
-        <dependency>
             <groupId>io.dropwizard.metrics</groupId>
             <artifactId>metrics-jetty9</artifactId>
         </dependency>
         <dependency>
             <groupId>org.eclipse.jetty</groupId>
             <artifactId>jetty-server</artifactId>
+            <exclusions>
+                <exclusion>
+                    <groupId>javax.servlet</groupId>
+                    <artifactId>javax.servlet-api</artifactId>
+                </exclusion>
+            </exclusions>
+        </dependency>
+        <dependency>
+            <groupId>jakarta.servlet</groupId>
+            <artifactId>jakarta.servlet-api</artifactId>
         </dependency>
         <dependency>
             <groupId>org.eclipse.jetty</groupId>
